chore: Added docs notice to README
I think it's *really* necessary here, as this is a mass of C code.
That's not anything easy to poke around in!

Harmful changes to the code will be forever accidental, resulting in
buffer overflows, weird bugs, and incorrect documentation. Let's work to
get this into a maintainable state - or RIIR ;)

also props to Benton for cleansing this of cpp lmao